move file around a bit more and rename. also add option to switch between old engine and new engine
moved some more files around diagnostic incremental analyzer and added option to choose between two engines.
