+++ src/Features/Core/Diagnostics/DiagnosticAnalyzerService_Test.cs
