+++ src/Features/Core/Diagnostics/DiagnosticAnalyzerService_UpdateSource.cs
