--- src/Features/Core/Diagnostics/EngineV1/DiagnosticAnalyzerService.IncrementalAnalyzer.AnalyzerExecutor.cs
@@ -1,182 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Immutable;
-using System.Diagnostics;
-using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.ErrorReporting;
-using Roslyn.Utilities;
-
-namespace Microsoft.CodeAnalysis.Diagnostics
-{
-    using ProviderId = Int32;
-
-    internal partial class DiagnosticAnalyzerService
-    {
-        internal partial class DiagnosticIncrementalAnalyzer
-        {
-            /// <summary>
-            /// AnalyzerExecutor returns analyzed data without side effect. 
-            /// it might uses cache to skip unnecessary duplicate analysis, but it will never update any internal state such as state
-            /// 
-            /// this is not finished form. as refactoring going on, this class will become more stateless in respect to caller, and less dependency on owner.
-            /// </summary>
-            internal class AnalyzerExecutor
-            {
-                private readonly DiagnosticIncrementalAnalyzer _owner;
-
-                public AnalyzerExecutor(DiagnosticIncrementalAnalyzer owner)
-                {
-                    _owner = owner;
-                }
-
-                public async Task<AnalysisData> GetSyntaxAnalysisDataAsync(
-                    DiagnosticAnalyzer provider, ProviderId providerId, VersionArgument versions, DiagnosticAnalyzerDriver analyzerDriver)
-                {
-                    try
-                    {
-                        var document = analyzerDriver.Document;
-                        var cancellationToken = analyzerDriver.CancellationToken;
-
-                        var state = AnalyzersAndState.GetOrCreateDiagnosticState(StateType.Syntax, providerId, provider, document.Project.Id, document.Project.Language);
-                        var existingData = await state.TryGetExistingDataAsync(document, cancellationToken).ConfigureAwait(false);
-
-                        if (CheckSyntaxVersions(document, existingData, versions))
-                        {
-                            return existingData;
-                        }
-
-                        var diagnosticData = await GetSyntaxDiagnosticsAsync(providerId, provider, analyzerDriver).ConfigureAwait(false);
-                        return new AnalysisData(versions.TextVersion, versions.DataVersion, GetExistingItems(existingData), diagnosticData.AsImmutableOrEmpty());
-                    }
-                    catch (Exception e) when(FatalError.ReportUnlessCanceled(e))
-                    {
-                        throw ExceptionUtilities.Unreachable;
-                    }
-                    }
-
-                public async Task<AnalysisData> GetDocumentAnalysisDataAsync(
-                    DiagnosticAnalyzer provider, ProviderId providerId, VersionArgument versions, DiagnosticAnalyzerDriver analyzerDrvier)
-                {
-                    try
-                    {
-                        var document = analyzerDrvier.Document;
-                        var cancellationToken = analyzerDrvier.CancellationToken;
-
-                        var state = AnalyzersAndState.GetOrCreateDiagnosticState(StateType.Document, providerId, provider, document.Project.Id, document.Project.Language);
-                        var existingData = await state.TryGetExistingDataAsync(document, cancellationToken).ConfigureAwait(false);
-
-                        if (CheckSemanticVersions(document, existingData, versions))
-                        {
-                            return existingData;
-                        }
-
-                        var diagnosticData = await GetSemanticDiagnosticsAsync(providerId, provider, analyzerDrvier).ConfigureAwait(false);
-                        return new AnalysisData(versions.TextVersion, versions.DataVersion, GetExistingItems(existingData), diagnosticData.AsImmutableOrEmpty());
-                    }
-                    catch (Exception e) when(FatalError.ReportUnlessCanceled(e))
-                    {
-                        throw ExceptionUtilities.Unreachable;
-                    }
-                    }
-
-                public async Task<AnalysisData> GetDocumentBodyAnalysisDataAsync(
-                    DiagnosticAnalyzer provider, ProviderId providerId, VersionArgument versions, DiagnosticAnalyzerDriver analyzerDriver,
-                    SyntaxNode root, SyntaxNode member, int memberId, bool supportsSemanticInSpan, MemberRangeMap.MemberRanges ranges)
-                {
-                    try
-                    {
-                        var document = analyzerDriver.Document;
-                        var cancellationToken = analyzerDriver.CancellationToken;
-
-                        var state = AnalyzersAndState.GetOrCreateDiagnosticState(StateType.Document, providerId, provider, document.Project.Id, document.Project.Language);
-                        var existingData = await state.TryGetExistingDataAsync(document, cancellationToken).ConfigureAwait(false);
-
-                        ImmutableArray<DiagnosticData> diagnosticData;
-                        if (supportsSemanticInSpan && CanUseDocumentState(existingData, ranges.TextVersion, versions.DataVersion))
-                        {
-                            var memberDxData = await GetSemanticDiagnosticsAsync(providerId, provider, analyzerDriver).ConfigureAwait(false);
-
-                            diagnosticData = _owner.UpdateDocumentDiagnostics(existingData, ranges.Ranges, memberDxData.AsImmutableOrEmpty(), root.SyntaxTree, member, memberId);
-                            ValidateMemberDiagnostics(providerId, provider, document, root, diagnosticData);
-                        }
-                        else
-                        {
-                            // if we can't re-use existing document state, only option we have is updating whole document state here.
-                            var dx = await GetSemanticDiagnosticsAsync(providerId, provider, analyzerDriver).ConfigureAwait(false);
-                            diagnosticData = dx.AsImmutableOrEmpty();
-                        }
-
-                        return new AnalysisData(versions.TextVersion, versions.DataVersion, GetExistingItems(existingData), diagnosticData);
-                    }
-                    catch (Exception e) when(FatalError.ReportUnlessCanceled(e))
-                    {
-                        throw ExceptionUtilities.Unreachable;
-                    }
-                    }
-
-                public async Task<AnalysisData> GetProjectAnalysisDataAsync(
-                    DiagnosticAnalyzer provider, ProviderId providerId, VersionArgument versions, DiagnosticAnalyzerDriver analyzerDriver)
-                {
-                    try
-                    {
-                        var project = analyzerDriver.Project;
-                        var cancellationToken = analyzerDriver.CancellationToken;
-
-                        var state = AnalyzersAndState.GetOrCreateDiagnosticState(StateType.Project, providerId, provider, project.Id, project.Language);
-                        var existingData = await state.TryGetExistingDataAsync(project, cancellationToken).ConfigureAwait(false);
-
-                        if (CheckSemanticVersions(project, existingData, versions))
-                        {
-                            return existingData;
-                        }
-
-                        // TODO: remove ForceAnalyzeAllDocuments at some point
-                        var diagnosticData = await GetProjectDiagnosticsAsync(providerId, provider, analyzerDriver, _owner.ForceAnalyzeAllDocuments).ConfigureAwait(false);
-                        return new AnalysisData(VersionStamp.Default, versions.DataVersion, GetExistingItems(existingData), diagnosticData.AsImmutableOrEmpty());
-                    }
-                    catch (Exception e) when(FatalError.ReportUnlessCanceled(e))
-                    {
-                        throw ExceptionUtilities.Unreachable;
-                    }
-                    }
-
-                private DiagnosticAnalyzersAndStates AnalyzersAndState
-                {
-                    get { return _owner._analyzersAndState; }
-                }
-
-                private bool CanUseDocumentState(AnalysisData existingData, VersionStamp textVersion, VersionStamp dataVersion)
-                {
-                    if (existingData == null)
-                    {
-                        return false;
-                    }
-
-                    // make sure data stored in the cache is one from its previous text update
-                    return existingData.DataVersion.Equals(dataVersion) && existingData.TextVersion.Equals(textVersion);
-                }
-
-                private static ImmutableArray<DiagnosticData> GetExistingItems(AnalysisData existingData)
-                {
-                    if (existingData == null)
-                    {
-                        return ImmutableArray<DiagnosticData>.Empty;
-                    }
-
-                    return existingData.Items;
-                }
-
-                [Conditional("DEBUG")]
-                private void ValidateMemberDiagnostics(ProviderId providerId, DiagnosticAnalyzer provider, Document document, SyntaxNode root, ImmutableArray<DiagnosticData> diagnostics)
-                {
-#if RANGE
-                var documentBasedDriver = new DiagnosticAnalyzerDriver(document, root.FullSpan, root, CancellationToken.None);
-                var expected = GetSemanticDiagnosticsAsync(providerId, provider, documentBasedDriver).WaitAndGetResult(documentBasedDriver.CancellationToken) ?? SpecializedCollections.EmptyEnumerable<DiagnosticData>();
-                Contract.Requires(diagnostics.SetEquals(expected));
-#endif
-                }
-            }
-        }
-    }
-}
