+++ src/Features/Core/Diagnostics/EngineV1/DiagnosticIncrementalAnalyzer.StateType.cs
@@ -2,7 +2,7 @@
 
 namespace Microsoft.CodeAnalysis.Diagnostics
 {
-    internal partial class DiagnosticAnalyzerService
+    internal partial class DiagnosticIncrementalAnalyzer
     {
         public enum StateType : int
         {
