+++ src/Features/Core/Features.csproj
@@ -191,14 +191,14 @@
     <Compile Include="Diagnostics\DiagnosticProviderMetadata.cs" />
     <Compile Include="Diagnostics\DiagnosticService.cs" />
     <Compile Include="Diagnostics\DiagnosticsUpdatedArgs.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.IncrementalAnalyzer.AnalyzerExecutor.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.IncrementalAnalyzer_GetDiagnostics.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.NestedTypes.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.StateType.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService_IncrementalAnalyzer.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService_UpdateSource.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService_Test.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.WorkspaceAnalyzersAndStates.PerLanguageAnalyzersAndStates.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.IncrementalAnalyzer.AnalyzerExecutor.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer_GetDiagnostics.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.NestedTypes.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.StateType.cs" />
+    <Compile Include="Diagnostics\DiagnosticAnalyzerService_IncrementalAnalyzer.cs" />
+    <Compile Include="Diagnostics\DiagnosticAnalyzerService_UpdateSource.cs" />
+    <Compile Include="Diagnostics\DiagnosticAnalyzerService_Test.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.WorkspaceAnalyzersAndStates.PerLanguageAnalyzersAndStates.cs" />
     <Compile Include="Diagnostics\EngineV1\IAnalyzerDriverService.cs" />
     <Compile Include="Diagnostics\IBuiltInAnalyzer.cs" />
     <Compile Include="Diagnostics\IDiagnosticAnalyzerService.cs" />
@@ -243,11 +243,11 @@
     <Compile Include="Diagnostics\Analyzers\SimplifyTypeNamesDiagnosticAnalyzerBase.cs" />
     <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerDriver.cs" />
     <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerCategory.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.DiagnosticAnalyzersAndStates.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.DiagnosticState.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.IncrementalAnalyzer.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.ProjectAnalyzersAndStates.cs" />
-    <Compile Include="Diagnostics\EngineV1\DiagnosticAnalyzerService.WorkspaceAnalyzersAndStates.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.DiagnosticAnalyzersAndStates.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.DiagnosticState.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.ProjectAnalyzersAndStates.cs" />
+    <Compile Include="Diagnostics\EngineV1\DiagnosticIncrementalAnalyzer.WorkspaceAnalyzersAndStates.cs" />
     <Compile Include="Diagnostics\EngineV1\ISyntaxNodeAnalyzerService.cs" />
     <Compile Include="Diagnostics\EngineV1\MemberRangeMap.cs" />
     <Compile Include="DocumentationCommentFormatting\AbstractDocumentationCommentFormattingService.cs" />
