+++ src/Features/Core/SolutionCrawler/SolutionCrawlerLogger.cs
@@ -247,7 +247,7 @@ namespace Microsoft.CodeAnalysis.SolutionCrawler
 
             foreach (var analyzer in analyzers)
             {
-                var diagIncrementalAnalyzer = analyzer as DiagnosticAnalyzerService.DiagnosticIncrementalAnalyzer;
+                var diagIncrementalAnalyzer = analyzer as DiagnosticIncrementalAnalyzer;
                 if (diagIncrementalAnalyzer != null)
                 {
                     diagIncrementalAnalyzer.LogAnalyzerCountSummary();
