+++ src/Features/Core/SolutionCrawler/WorkCoordinator.IncrementalAnalyzerProcessor.cs
@@ -80,8 +80,8 @@ namespace Microsoft.CodeAnalysis.SolutionCrawler
 
                 private static ImmutableArray<IIncrementalAnalyzer> OrderAnalyzers(IEnumerable<IIncrementalAnalyzer> analyzers)
                 {
-                    return SpecializedCollections.SingletonEnumerable(analyzers.FirstOrDefault(a => a.GetType() == typeof(DiagnosticAnalyzerService.DiagnosticIncrementalAnalyzer)))
-                                                                              .Concat(analyzers.Where(a => a.GetType() != typeof(DiagnosticAnalyzerService.DiagnosticIncrementalAnalyzer)))
+                    return SpecializedCollections.SingletonEnumerable(analyzers.FirstOrDefault(a => a.GetType() == typeof(DiagnosticIncrementalAnalyzer)))
+                                                                              .Concat(analyzers.Where(a => a.GetType() != typeof(DiagnosticIncrementalAnalyzer)))
                                                                               .WhereNotNull().ToImmutableArray();
                 }
 
@@ -204,11 +204,11 @@ namespace Microsoft.CodeAnalysis.SolutionCrawler
                         // re-run just the body
                         await RunAnalyzersAsync(analyzers, document, (a, d, c) => a.AnalyzeDocumentAsync(d, activeMember, c), cancellationToken).ConfigureAwait(false);
                     }
-                    catch (Exception e) when(FatalError.ReportUnlessCanceled(e))
+                    catch (Exception e) when (FatalError.ReportUnlessCanceled(e))
                     {
                         throw ExceptionUtilities.Unreachable;
                     }
-                    }
+                }
 
                 private static async Task<TResult> GetOrDefaultAsync<TData, TResult>(TData value, Func<TData, CancellationToken, Task<TResult>> funcAsync, CancellationToken cancellationToken)
                 {
@@ -220,16 +220,16 @@ namespace Microsoft.CodeAnalysis.SolutionCrawler
                     {
                         return default(TResult);
                     }
-                    catch (AggregateException e) when(CrashUnlessCanceled(e))
+                    catch (AggregateException e) when (CrashUnlessCanceled(e))
                     {
                         return default(TResult);
                     }
-                    catch (Exception e) when(FatalError.Report(e))
+                    catch (Exception e) when (FatalError.Report(e))
                     {
                         // TODO: manage bad workers like what code actions does now
                         throw ExceptionUtilities.Unreachable;
                     }
-                    }
+                }
 
                 private static SyntaxNode GetMemberNode(ISyntaxFactsService service, SyntaxNode root, SyntaxPath memberPath)
                 {
